Store the sampling rate as a float value.

+++ lib/psysmon/packages/obspyImportWaveform/__init__.py
@@ -58,7 +58,7 @@ def databaseFactory(base):
         network = Column(String(10), nullable=False, default='')
         recorder_serial = Column(String(45), nullable=False)
         stream = Column(String(45), nullable=False)
-        sps = Column(Integer, nullable=False)
+        sps = Column(Float(53), nullable=False)
         numsamp = Column(Integer, nullable=False)
         begin_date = Column(String(26), nullable=False)
         begin_time = Column(Float(53), nullable=False)
